all UI current
current UI, medicine data xlsx is for holding info (do not change rn), same with user manual. medicine data UI is as it was during last demo, likely won't be updated to new version till after design review but i will work on a new version of it with what nick and sam said. i'd reccomend making a copy of it so the code isn't all totally changed from the original.